requirements.txt
python-dotenv
sentence-transformers
requests
urlparser
pandas
elasticsearch
asyncio
pymystem3
pydantic
aiohttp
uvicorn
fastapi